Remove unused components
