+++ screens/FoundDogScreen.js
@@ -1,19 +1,11 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import PropTypes from 'prop-types'
-import {
-  View,
-  StyleSheet,
-  Image,
-  TouchableOpacity,
-  SafeAreaView,
-  ScrollView,
-} from 'react-native'
+import { StyleSheet, SafeAreaView } from 'react-native'
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view'
-import FoundDogForm from '../components/FoundDogForm/FoundDogForm2'
+import FoundDogForm from '../components/FoundDogForm/FoundDogForm'
 import { postFoundDog } from '../redux/actions/foundDog'
 import { getSignedUrl } from '../redux/actions/images'
-import { Header } from 'react-native-elements'
 
 class FoundDogScreen extends Component {
   constructor(props) {
@@ -23,7 +15,6 @@ class FoundDogScreen extends Component {
     this.imagePath = this.props.navigation.getParam('uri')
     this.pressPicture = this.pressPicture.bind(this)
   }
-  async componentDidMount() {}
 
   async onSubmitHandler(token, data) {
     const imageLink = await this.uploadImage(this.imagePath)
